refactor: outsource the definitions for the meta statics from the main statistics module
