+++ src/statistics/index.ts
@@ -2,3 +2,5 @@ export * from './statistics'
 export * from './features'
 export * from './output'
 export * from './post-process'
+export { initialMetaStatistics } from './meta-statistics'
+export { MetaStatistics } from './meta-statistics'
